Fix bug where videos would forever be re-added to the task list if force transcoding was enabled.
Added a flag that will be set if a file was transcoded with the force setting enabled.
This flag will prevent the file being transcoded again in the future in a loop.
This flag will not, however, prevent the plugin from transcoding again in the future if a smart filter is configured and the plugin detects that this filter applies to the scanned file.